% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycifr.R
\name{gaussian_fit}
\alias{gaussian_fit}
\title{Fit a Gaussian Mixture Model to a vector of data}
\usage{
gaussian_fit(
  x,
  lower.limit = 0.001,
  upper.limit = 0.999,
  num.gaussian = 2,
  plot = TRUE
)
}
\arguments{
\item{x}{A numeric vector of data to fit}

\item{lower.limit}{A numeric value between 0 and 1 specifying the lower limit for data points to include in the fit (default: 0.001)}

\item{upper.limit}{A numeric value between 0 and 1 specifying the upper limit for data points to include in the fit (default: 0.999)}

\item{num.gaussian}{An integer specifying the number of Gaussians to fit (default: 2)}

\item{plot}{A logical value indicating whether to generate a histogram with fitted Gaussians and threshold (default: TRUE)}
}
\value{
A data.table with the model parameters, including the means and variances of the fitted Gaussians.
}
\description{
This function fits a Gaussian Mixture Model to a vector of data using the Mclust package in R. It returns a data.table with the model parameters, including the means and variances of the fitted Gaussians. Optionally, it also generates a histogram of the data with lines for the fitted Gaussians and a vertical line at the threshold between the two Gaussians.
}
\examples{
data <- rnorm(100)
gaussian_fit(data)

}
